Use of Openstack CI for your own projects doc
OpenStack CI is a great system and useful for properly bringing
a CI workflow into any project, not only OpenStack.

This presentation will cover the basics about the CI process
we follow at Openstack. It will dig on how it can be reused to setup
a great CI system for own projects. It will be showing the basics of
deployment and configuration, highlighting the efforts Infra team
is doing to make the CI system totally independent from
Openstack requirements.

Change-Id: I157b40f96fb70e638235c6014235694dd5462e38