Add README.rst to using-your-own slide deck
Turns out, our make-index reads the slide deck title from the
README.rst, so if it doesn't exist, or uses the default one that
comes with the template, the talk title is incorrect on the
resulting publications site.

Text for this README adapted from the talk description at Fossetcon,
where this talk was given in September 2014.

Change-Id: Iefed788430553f2e117476f605d2f6ee41c89a10